implemented authentication and email verification using firebase and firestore

+++ server/src/config/firebase.config.js
@@ -1,6 +1,7 @@
 import env from './env.config.js';
 import { initializeApp } from 'firebase/app';
 import { collection, getFirestore, addDoc, setDoc, doc, getDoc, getDocs, deleteDoc, updateDoc } from 'firebase/firestore';
+import admin from 'firebase-admin';
 
 // Firebase configuration
 const firebaseConfig = {
@@ -18,9 +19,15 @@ console.log('Firebase Config Check:', {
     apiKey: firebaseConfig.apiKey ? 'Set' : 'Missing'
 });
 
+if (!admin.apps.length) {
+  admin.initializeApp({
+    credential: admin.credential.cert('./src/config/keyfile.json'),
+  });
+}
+
 const app = initializeApp(firebaseConfig);
 const db = getFirestore(app);
 const users = collection(db, 'users');
 const mail = collection(db, 'mail');
 
-export { app, db, users, mail, addDoc, setDoc, doc, getDoc, getDocs, deleteDoc, updateDoc };
+export { app,admin, db, users, mail, addDoc, setDoc, doc, getDoc, getDocs, deleteDoc, updateDoc };
